functions/scripts/twint-client/requirements.txt
# pip3 install -r requirements.txt
requests
python-firebase
google-cloud-firestore
twint